DSP: separated resampler to its own file and improvements on several DSP methods. DSP: Much improved sinc resampler. Added some intrinsics. added start of soxr implementation. WireDlg: More improvements in the visuals: RMS and PEAK in VU, nicer osci, improvements in spectrum, and some fixes in phase. Sampulse/Sampler: Added support for samplerate, changed finetune range to 100 cents and other related changes. Also samplerate support in xmsongexport. Sampulse: Properly working .iti (v1 and v2) loaders. Dialogs : Improvements in sampler and sampulse dialogs (both, visual and usability). Other miscellaneous source changes (arrays made multiples of 4, some casts and type usages, and better definition of AudioDriverInfo)

+++ trunk/psycle/src/psycle/host/InfoDlg.cpp
@@ -90,7 +90,7 @@ namespace psycle { namespace host {
 				CSingleLock lock(&Global::song().semaphore, FALSE);
 				if (!lock.Lock(50)) return;
 
-				unsigned long num_threads_running = Global::player().num_threads();
+				size_t num_threads_running = Global::player().num_threads();
 				char buffer[128];
 				
 				cpu_time_clock::duration total_machine_processing_time(0);
